Убрали все фигуры в модуль, по одному классу на файл.
В архитектуре явная ошибка: доска по умолчанию знает о фигурах, фигуры знают о доске. Получается, что если вместо шахмат нам нужны будут шашки, то все придется переделывать с нуля. Как развести ответственность, пока не очень понимаю.
