+++ src/domain/figures/__init__.py
@@ -1,0 +1,12 @@
+# pylint:disable=import-error, cyclic-import
+"""
+Пакет шахматных фигур
+"""
+
+from .figure import Figure
+from .pawn import Pawn
+from .rook import Rook
+from .bishop import Bishop
+from .knight import Knight
+from .queen import Queen
+from .king import King
